feat: add sensitive information check
This commit adds a check for sensitive information in staged files before committing. If any sensitive information is found, the commit will be aborted and a warning message will be displayed.

The following types of sensitive information are currently checked for:

- API keys
- Private keys
- Passwords
- Secrets
- Access tokens
- Private SSH keys

This check helps to prevent sensitive information from being accidentally committed and exposed to unauthorized users.